requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bitarray==3.3.1
canonicaljson==2.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
ckzg==2.1.1
cytoolz==1.0.1
distro==1.9.0
ecdsa==0.19.1
eth-account==0.13.6
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
farcaster==0.7.12
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
h11==0.14.0
hexbytes==1.3.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
jiter==0.9.0
multidict==6.4.3
numpy==2.2.5
openai==0.28.0
outcome==1.3.0.post0
packaging==25.0
parsimonious==0.10.0
praw==7.8.1
prawcore==2.4.0
propcache==0.3.1
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
pyhumps==3.8.0
PySocks==1.7.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
safetensors==0.5.3
schedule==1.2.2
selenium==4.31.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tokenizers==0.21.1
toolz==1.0.0
tqdm==4.67.1
transformers==4.51.3
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
update-checker==0.18.0
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.20.0
